fantasy-cricket-analyzer/requirements.txt
CrewAI
pandas
pyyaml
streamlit